+++ src/aks-preview/azext_aks_preview/vendored_sdks/azure_mgmt_preview_aks/_container_service_client.py
@@ -53,7 +53,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
     :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
     """
 
-    DEFAULT_API_VERSION = '2023-04-01'
+    DEFAULT_API_VERSION = '2023-06-01'
     _PROFILE_TAG = "azure.mgmt.containerservice.ContainerServiceClient"
     LATEST_PROFILE = ProfileDefinition({
         _PROFILE_TAG: {
@@ -75,6 +75,8 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         profile: KnownProfiles=KnownProfiles.default,
         **kwargs: Any
     ):
+        if api_version:
+            kwargs.setdefault('api_version', api_version)
         self._config = ContainerServiceClientConfiguration(credential, subscription_id, **kwargs)
         self._client = ARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
         super(ContainerServiceClient, self).__init__(
@@ -148,7 +150,10 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-03-02-preview: :mod:`v2023_03_02_preview.models<azure.mgmt.containerservice.v2023_03_02_preview.models>`
            * 2023-04-01: :mod:`v2023_04_01.models<azure.mgmt.containerservice.v2023_04_01.models>`
            * 2023-04-02-preview: :mod:`v2023_04_02_preview.models<azure.mgmt.containerservice.v2023_04_02_preview.models>`
+           * 2023-05-01: :mod:`v2023_05_01.models<azure.mgmt.containerservice.v2023_05_01.models>`
            * 2023-05-02-preview: :mod:`v2023_05_02_preview.models<azure.mgmt.containerservice.v2023_05_02_preview.models>`
+           * 2023-06-01: :mod:`v2023_06_01.models<azure.mgmt.containerservice.v2023_06_01.models>`
+           * 2023-06-02-preview: :mod:`v2023_06_02_preview.models<azure.mgmt.containerservice.v2023_06_02_preview.models>`
         """
         if api_version == '2017-07-01':
             from .v2017_07_01 import models
@@ -324,8 +329,17 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         elif api_version == '2023-04-02-preview':
             from .v2023_04_02_preview import models
             return models
+        elif api_version == '2023-05-01':
+            from .v2023_05_01 import models
+            return models
         elif api_version == '2023-05-02-preview':
             from .v2023_05_02_preview import models
+            return models
+        elif api_version == '2023-06-01':
+            from .v2023_06_01 import models
+            return models
+        elif api_version == '2023-06-02-preview':
+            from .v2023_06_02_preview import models
             return models
         raise ValueError("API version {} is not available".format(api_version))
 
@@ -384,7 +398,10 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-03-02-preview: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2023_03_02_preview.operations.AgentPoolsOperations>`
            * 2023-04-01: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2023_04_01.operations.AgentPoolsOperations>`
            * 2023-04-02-preview: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2023_04_02_preview.operations.AgentPoolsOperations>`
+           * 2023-05-01: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2023_05_01.operations.AgentPoolsOperations>`
            * 2023-05-02-preview: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2023_05_02_preview.operations.AgentPoolsOperations>`
+           * 2023-06-01: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2023_06_01.operations.AgentPoolsOperations>`
+           * 2023-06-02-preview: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2023_06_02_preview.operations.AgentPoolsOperations>`
         """
         api_version = self._get_api_version('agent_pools')
         if api_version == '2019-02-01':
@@ -489,12 +506,18 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_04_01.operations import AgentPoolsOperations as OperationClass
         elif api_version == '2023-04-02-preview':
             from .v2023_04_02_preview.operations import AgentPoolsOperations as OperationClass
+        elif api_version == '2023-05-01':
+            from .v2023_05_01.operations import AgentPoolsOperations as OperationClass
         elif api_version == '2023-05-02-preview':
             from .v2023_05_02_preview.operations import AgentPoolsOperations as OperationClass
+        elif api_version == '2023-06-01':
+            from .v2023_06_01.operations import AgentPoolsOperations as OperationClass
+        elif api_version == '2023-06-02-preview':
+            from .v2023_06_02_preview.operations import AgentPoolsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'agent_pools'".format(api_version))
         self._config.api_version = api_version
-        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
     @property
     def container_services(self):
@@ -511,7 +534,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         else:
             raise ValueError("API version {} does not have operation group 'container_services'".format(api_version))
         self._config.api_version = api_version
-        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
     @property
     def fleet_members(self):
@@ -531,7 +554,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         else:
             raise ValueError("API version {} does not have operation group 'fleet_members'".format(api_version))
         self._config.api_version = api_version
-        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
     @property
     def fleets(self):
@@ -551,7 +574,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         else:
             raise ValueError("API version {} does not have operation group 'fleets'".format(api_version))
         self._config.api_version = api_version
-        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
     @property
     def maintenance_configurations(self):
@@ -594,7 +617,10 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-03-02-preview: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2023_03_02_preview.operations.MaintenanceConfigurationsOperations>`
            * 2023-04-01: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2023_04_01.operations.MaintenanceConfigurationsOperations>`
            * 2023-04-02-preview: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2023_04_02_preview.operations.MaintenanceConfigurationsOperations>`
+           * 2023-05-01: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2023_05_01.operations.MaintenanceConfigurationsOperations>`
            * 2023-05-02-preview: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2023_05_02_preview.operations.MaintenanceConfigurationsOperations>`
+           * 2023-06-01: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2023_06_01.operations.MaintenanceConfigurationsOperations>`
+           * 2023-06-02-preview: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2023_06_02_preview.operations.MaintenanceConfigurationsOperations>`
         """
         api_version = self._get_api_version('maintenance_configurations')
         if api_version == '2020-12-01':
@@ -671,12 +697,18 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_04_01.operations import MaintenanceConfigurationsOperations as OperationClass
         elif api_version == '2023-04-02-preview':
             from .v2023_04_02_preview.operations import MaintenanceConfigurationsOperations as OperationClass
+        elif api_version == '2023-05-01':
+            from .v2023_05_01.operations import MaintenanceConfigurationsOperations as OperationClass
         elif api_version == '2023-05-02-preview':
             from .v2023_05_02_preview.operations import MaintenanceConfigurationsOperations as OperationClass
+        elif api_version == '2023-06-01':
+            from .v2023_06_01.operations import MaintenanceConfigurationsOperations as OperationClass
+        elif api_version == '2023-06-02-preview':
+            from .v2023_06_02_preview.operations import MaintenanceConfigurationsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'maintenance_configurations'".format(api_version))
         self._config.api_version = api_version
-        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
     @property
     def managed_cluster_snapshots(self):
@@ -698,6 +730,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-03-02-preview: :class:`ManagedClusterSnapshotsOperations<azure.mgmt.containerservice.v2023_03_02_preview.operations.ManagedClusterSnapshotsOperations>`
            * 2023-04-02-preview: :class:`ManagedClusterSnapshotsOperations<azure.mgmt.containerservice.v2023_04_02_preview.operations.ManagedClusterSnapshotsOperations>`
            * 2023-05-02-preview: :class:`ManagedClusterSnapshotsOperations<azure.mgmt.containerservice.v2023_05_02_preview.operations.ManagedClusterSnapshotsOperations>`
+           * 2023-06-02-preview: :class:`ManagedClusterSnapshotsOperations<azure.mgmt.containerservice.v2023_06_02_preview.operations.ManagedClusterSnapshotsOperations>`
         """
         api_version = self._get_api_version('managed_cluster_snapshots')
         if api_version == '2022-02-02-preview':
@@ -732,10 +765,12 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_04_02_preview.operations import ManagedClusterSnapshotsOperations as OperationClass
         elif api_version == '2023-05-02-preview':
             from .v2023_05_02_preview.operations import ManagedClusterSnapshotsOperations as OperationClass
+        elif api_version == '2023-06-02-preview':
+            from .v2023_06_02_preview.operations import ManagedClusterSnapshotsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'managed_cluster_snapshots'".format(api_version))
         self._config.api_version = api_version
-        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
     @property
     def managed_clusters(self):
@@ -794,7 +829,10 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-03-02-preview: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2023_03_02_preview.operations.ManagedClustersOperations>`
            * 2023-04-01: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2023_04_01.operations.ManagedClustersOperations>`
            * 2023-04-02-preview: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2023_04_02_preview.operations.ManagedClustersOperations>`
+           * 2023-05-01: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2023_05_01.operations.ManagedClustersOperations>`
            * 2023-05-02-preview: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2023_05_02_preview.operations.ManagedClustersOperations>`
+           * 2023-06-01: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2023_06_01.operations.ManagedClustersOperations>`
+           * 2023-06-02-preview: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2023_06_02_preview.operations.ManagedClustersOperations>`
         """
         api_version = self._get_api_version('managed_clusters')
         if api_version == '2018-03-31':
@@ -903,12 +941,18 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_04_01.operations import ManagedClustersOperations as OperationClass
         elif api_version == '2023-04-02-preview':
             from .v2023_04_02_preview.operations import ManagedClustersOperations as OperationClass
+        elif api_version == '2023-05-01':
+            from .v2023_05_01.operations import ManagedClustersOperations as OperationClass
         elif api_version == '2023-05-02-preview':
             from .v2023_05_02_preview.operations import ManagedClustersOperations as OperationClass
+        elif api_version == '2023-06-01':
+            from .v2023_06_01.operations import ManagedClustersOperations as OperationClass
+        elif api_version == '2023-06-02-preview':
+            from .v2023_06_02_preview.operations import ManagedClustersOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'managed_clusters'".format(api_version))
         self._config.api_version = api_version
-        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
     @property
     def open_shift_managed_clusters(self):
@@ -931,7 +975,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         else:
             raise ValueError("API version {} does not have operation group 'open_shift_managed_clusters'".format(api_version))
         self._config.api_version = api_version
-        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
     @property
     def operations(self):
@@ -990,7 +1034,10 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-03-02-preview: :class:`Operations<azure.mgmt.containerservice.v2023_03_02_preview.operations.Operations>`
            * 2023-04-01: :class:`Operations<azure.mgmt.containerservice.v2023_04_01.operations.Operations>`
            * 2023-04-02-preview: :class:`Operations<azure.mgmt.containerservice.v2023_04_02_preview.operations.Operations>`
+           * 2023-05-01: :class:`Operations<azure.mgmt.containerservice.v2023_05_01.operations.Operations>`
            * 2023-05-02-preview: :class:`Operations<azure.mgmt.containerservice.v2023_05_02_preview.operations.Operations>`
+           * 2023-06-01: :class:`Operations<azure.mgmt.containerservice.v2023_06_01.operations.Operations>`
+           * 2023-06-02-preview: :class:`Operations<azure.mgmt.containerservice.v2023_06_02_preview.operations.Operations>`
         """
         api_version = self._get_api_version('operations')
         if api_version == '2018-03-31':
@@ -1099,12 +1146,18 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_04_01.operations import Operations as OperationClass
         elif api_version == '2023-04-02-preview':
             from .v2023_04_02_preview.operations import Operations as OperationClass
+        elif api_version == '2023-05-01':
+            from .v2023_05_01.operations import Operations as OperationClass
         elif api_version == '2023-05-02-preview':
             from .v2023_05_02_preview.operations import Operations as OperationClass
+        elif api_version == '2023-06-01':
+            from .v2023_06_01.operations import Operations as OperationClass
+        elif api_version == '2023-06-02-preview':
+            from .v2023_06_02_preview.operations import Operations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'operations'".format(api_version))
         self._config.api_version = api_version
-        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
     @property
     def private_endpoint_connections(self):
@@ -1151,7 +1204,10 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-03-02-preview: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2023_03_02_preview.operations.PrivateEndpointConnectionsOperations>`
            * 2023-04-01: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2023_04_01.operations.PrivateEndpointConnectionsOperations>`
            * 2023-04-02-preview: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2023_04_02_preview.operations.PrivateEndpointConnectionsOperations>`
+           * 2023-05-01: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2023_05_01.operations.PrivateEndpointConnectionsOperations>`
            * 2023-05-02-preview: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2023_05_02_preview.operations.PrivateEndpointConnectionsOperations>`
+           * 2023-06-01: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2023_06_01.operations.PrivateEndpointConnectionsOperations>`
+           * 2023-06-02-preview: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2023_06_02_preview.operations.PrivateEndpointConnectionsOperations>`
         """
         api_version = self._get_api_version('private_endpoint_connections')
         if api_version == '2020-06-01':
@@ -1236,12 +1292,18 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_04_01.operations import PrivateEndpointConnectionsOperations as OperationClass
         elif api_version == '2023-04-02-preview':
             from .v2023_04_02_preview.operations import PrivateEndpointConnectionsOperations as OperationClass
+        elif api_version == '2023-05-01':
+            from .v2023_05_01.operations import PrivateEndpointConnectionsOperations as OperationClass
         elif api_version == '2023-05-02-preview':
             from .v2023_05_02_preview.operations import PrivateEndpointConnectionsOperations as OperationClass
+        elif api_version == '2023-06-01':
+            from .v2023_06_01.operations import PrivateEndpointConnectionsOperations as OperationClass
+        elif api_version == '2023-06-02-preview':
+            from .v2023_06_02_preview.operations import PrivateEndpointConnectionsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'private_endpoint_connections'".format(api_version))
         self._config.api_version = api_version
-        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
     @property
     def private_link_resources(self):
@@ -1286,7 +1348,10 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-03-02-preview: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2023_03_02_preview.operations.PrivateLinkResourcesOperations>`
            * 2023-04-01: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2023_04_01.operations.PrivateLinkResourcesOperations>`
            * 2023-04-02-preview: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2023_04_02_preview.operations.PrivateLinkResourcesOperations>`
+           * 2023-05-01: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2023_05_01.operations.PrivateLinkResourcesOperations>`
            * 2023-05-02-preview: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2023_05_02_preview.operations.PrivateLinkResourcesOperations>`
+           * 2023-06-01: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2023_06_01.operations.PrivateLinkResourcesOperations>`
+           * 2023-06-02-preview: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2023_06_02_preview.operations.PrivateLinkResourcesOperations>`
         """
         api_version = self._get_api_version('private_link_resources')
         if api_version == '2020-09-01':
@@ -1367,12 +1432,18 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_04_01.operations import PrivateLinkResourcesOperations as OperationClass
         elif api_version == '2023-04-02-preview':
             from .v2023_04_02_preview.operations import PrivateLinkResourcesOperations as OperationClass
+        elif api_version == '2023-05-01':
+            from .v2023_05_01.operations import PrivateLinkResourcesOperations as OperationClass
         elif api_version == '2023-05-02-preview':
             from .v2023_05_02_preview.operations import PrivateLinkResourcesOperations as OperationClass
+        elif api_version == '2023-06-01':
+            from .v2023_06_01.operations import PrivateLinkResourcesOperations as OperationClass
+        elif api_version == '2023-06-02-preview':
+            from .v2023_06_02_preview.operations import PrivateLinkResourcesOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'private_link_resources'".format(api_version))
         self._config.api_version = api_version
-        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
     @property
     def resolve_private_link_service_id(self):
@@ -1417,7 +1488,10 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-03-02-preview: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2023_03_02_preview.operations.ResolvePrivateLinkServiceIdOperations>`
            * 2023-04-01: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2023_04_01.operations.ResolvePrivateLinkServiceIdOperations>`
            * 2023-04-02-preview: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2023_04_02_preview.operations.ResolvePrivateLinkServiceIdOperations>`
+           * 2023-05-01: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2023_05_01.operations.ResolvePrivateLinkServiceIdOperations>`
            * 2023-05-02-preview: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2023_05_02_preview.operations.ResolvePrivateLinkServiceIdOperations>`
+           * 2023-06-01: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2023_06_01.operations.ResolvePrivateLinkServiceIdOperations>`
+           * 2023-06-02-preview: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2023_06_02_preview.operations.ResolvePrivateLinkServiceIdOperations>`
         """
         api_version = self._get_api_version('resolve_private_link_service_id')
         if api_version == '2020-09-01':
@@ -1498,12 +1572,18 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_04_01.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
         elif api_version == '2023-04-02-preview':
             from .v2023_04_02_preview.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
+        elif api_version == '2023-05-01':
+            from .v2023_05_01.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
         elif api_version == '2023-05-02-preview':
             from .v2023_05_02_preview.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
+        elif api_version == '2023-06-01':
+            from .v2023_06_01.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
+        elif api_version == '2023-06-02-preview':
+            from .v2023_06_02_preview.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'resolve_private_link_service_id'".format(api_version))
         self._config.api_version = api_version
-        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
     @property
     def snapshots(self):
@@ -1541,7 +1621,10 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-03-02-preview: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2023_03_02_preview.operations.SnapshotsOperations>`
            * 2023-04-01: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2023_04_01.operations.SnapshotsOperations>`
            * 2023-04-02-preview: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2023_04_02_preview.operations.SnapshotsOperations>`
+           * 2023-05-01: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2023_05_01.operations.SnapshotsOperations>`
            * 2023-05-02-preview: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2023_05_02_preview.operations.SnapshotsOperations>`
+           * 2023-06-01: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2023_06_01.operations.SnapshotsOperations>`
+           * 2023-06-02-preview: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2023_06_02_preview.operations.SnapshotsOperations>`
         """
         api_version = self._get_api_version('snapshots')
         if api_version == '2021-08-01':
@@ -1608,12 +1691,18 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_04_01.operations import SnapshotsOperations as OperationClass
         elif api_version == '2023-04-02-preview':
             from .v2023_04_02_preview.operations import SnapshotsOperations as OperationClass
+        elif api_version == '2023-05-01':
+            from .v2023_05_01.operations import SnapshotsOperations as OperationClass
         elif api_version == '2023-05-02-preview':
             from .v2023_05_02_preview.operations import SnapshotsOperations as OperationClass
+        elif api_version == '2023-06-01':
+            from .v2023_06_01.operations import SnapshotsOperations as OperationClass
+        elif api_version == '2023-06-02-preview':
+            from .v2023_06_02_preview.operations import SnapshotsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'snapshots'".format(api_version))
         self._config.api_version = api_version
-        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
     @property
     def trusted_access_role_bindings(self):
@@ -1633,6 +1722,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-03-02-preview: :class:`TrustedAccessRoleBindingsOperations<azure.mgmt.containerservice.v2023_03_02_preview.operations.TrustedAccessRoleBindingsOperations>`
            * 2023-04-02-preview: :class:`TrustedAccessRoleBindingsOperations<azure.mgmt.containerservice.v2023_04_02_preview.operations.TrustedAccessRoleBindingsOperations>`
            * 2023-05-02-preview: :class:`TrustedAccessRoleBindingsOperations<azure.mgmt.containerservice.v2023_05_02_preview.operations.TrustedAccessRoleBindingsOperations>`
+           * 2023-06-02-preview: :class:`TrustedAccessRoleBindingsOperations<azure.mgmt.containerservice.v2023_06_02_preview.operations.TrustedAccessRoleBindingsOperations>`
         """
         api_version = self._get_api_version('trusted_access_role_bindings')
         if api_version == '2022-04-02-preview':
@@ -1663,10 +1753,12 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_04_02_preview.operations import TrustedAccessRoleBindingsOperations as OperationClass
         elif api_version == '2023-05-02-preview':
             from .v2023_05_02_preview.operations import TrustedAccessRoleBindingsOperations as OperationClass
+        elif api_version == '2023-06-02-preview':
+            from .v2023_06_02_preview.operations import TrustedAccessRoleBindingsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'trusted_access_role_bindings'".format(api_version))
         self._config.api_version = api_version
-        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
     @property
     def trusted_access_roles(self):
@@ -1686,6 +1778,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-03-02-preview: :class:`TrustedAccessRolesOperations<azure.mgmt.containerservice.v2023_03_02_preview.operations.TrustedAccessRolesOperations>`
            * 2023-04-02-preview: :class:`TrustedAccessRolesOperations<azure.mgmt.containerservice.v2023_04_02_preview.operations.TrustedAccessRolesOperations>`
            * 2023-05-02-preview: :class:`TrustedAccessRolesOperations<azure.mgmt.containerservice.v2023_05_02_preview.operations.TrustedAccessRolesOperations>`
+           * 2023-06-02-preview: :class:`TrustedAccessRolesOperations<azure.mgmt.containerservice.v2023_06_02_preview.operations.TrustedAccessRolesOperations>`
         """
         api_version = self._get_api_version('trusted_access_roles')
         if api_version == '2022-04-02-preview':
@@ -1716,10 +1809,12 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_04_02_preview.operations import TrustedAccessRolesOperations as OperationClass
         elif api_version == '2023-05-02-preview':
             from .v2023_05_02_preview.operations import TrustedAccessRolesOperations as OperationClass
+        elif api_version == '2023-06-02-preview':
+            from .v2023_06_02_preview.operations import TrustedAccessRolesOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'trusted_access_roles'".format(api_version))
         self._config.api_version = api_version
-        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
     def close(self):
         self._client.close()