+++ src/devcenter/azext_devcenter/aaz/latest/devcenter/admin/devbox_definition/_update.py
@@ -48,7 +48,7 @@ class Update(AAZCommand):
 
         _args_schema = cls._args_schema
         _args_schema.dev_box_definition_name = AAZStrArg(
-            options=["-n", "--name", "--dev-box-definition-name"],
+            options=["-n", "--name", "--devbox-definition-name"],
             help="The name of the dev box definition.",
             required=True,
             id_part="child_name_1",