+++ src/managednetworkfabric/azext_managednetworkfabric/tests/latest/test_l2domain.py
@@ -25,8 +25,8 @@ def call_scenario1(test):
     setup_scenario1(test)
     step_create(test, checks=[])
     step_show(test, checks=[])
-    step_list_subscription(test, checks=[])
     step_list_resource_group(test, checks=[])
+    step_list_subscription(test, checks=[])
     step_update_admin_state_Enable(test, checks=[])
     step_update_admin_state_Disable(test, checks=[])
     step_delete(test, checks=[])
@@ -36,7 +36,7 @@ def step_create(test, checks=None):
     '''l2domain create operation'''
     if checks is None:
         checks = []
-    test.cmd('az networkfabric l2domain create --resource-group {rg} --resource-name {name} --location {location} --nf-id {nf_id} --vlan-id {vlan_id}', checks=checks)
+    test.cmd('az networkfabric l2domain create --resource-group {rg} --resource-name {name} --location {location} --nf-id {nf_id} --vlan-id {vlan_id} --mtu {mtu}', checks=checks)
 
 def step_show(test, checks=None):
     '''l2domain show operation'''
@@ -44,19 +44,6 @@ def step_show(test, checks=None):
         checks = []
     test.cmd(
         'az networkfabric l2domain show --resource-name {name} --resource-group {rg}')
-    
-def step_delete(test, checks=None):
-    '''l2domain delete operation'''
-    if checks is None:
-        checks = []
-    test.cmd(
-        'az networkfabric l2domain delete --resource-name {name} --resource-group {rg}')
-
-def step_list_subscription(test, checks=None):
-    '''l2domain list by subscription operation'''
-    if checks is None:
-        checks = []
-    test.cmd('az networkfabric l2domain list')
 
 def step_list_resource_group(test, checks=None):
     '''l2domain list by resource group operation'''
@@ -64,19 +51,32 @@ def step_list_resource_group(test, checks=None):
         checks = []
     test.cmd('az networkfabric l2domain list --resource-group {rg}')
 
+def step_list_subscription(test, checks=None):
+    '''l2domain list by subscription operation'''
+    if checks is None:
+        checks = []
+    test.cmd('az networkfabric l2domain list')
+
 def step_update_admin_state_Enable(test, checks=None):
     '''l2domain Update admin state operation'''
     if checks is None:
         checks = []
-    test.cmd('az networkfabric l2domain update-admin-state --resource-group {rg} --resource-name {name} --state {state_Enable} --no-wait')
+    test.cmd('az networkfabric l2domain update-admin-state --resource-group {rg} --resource-name {name} --state {state_Enable}')
 
 def step_update_admin_state_Disable(test, checks=None):
     '''l2domain Update admin state operation'''
     if checks is None:
         checks = []
-    test.cmd('az networkfabric l2domain update-admin-state --resource-group {rg} --resource-name {name} --state {state_Disable} --no-wait')
+    test.cmd('az networkfabric l2domain update-admin-state --resource-group {rg} --resource-name {name} --state {state_Disable}')
 
-class L2DomainScenarioTest1(ScenarioTest):
+def step_delete(test, checks=None):
+    '''l2domain delete operation'''
+    if checks is None:
+        checks = []
+    test.cmd(
+        'az networkfabric l2domain delete --resource-name {name} --resource-group {rg}')
+
+class GA_L2DomainScenarioTest1(ScenarioTest):
     ''' L2 Domain Scenario test'''
 
     def __init__(self, *args, **kwargs):
@@ -92,6 +92,6 @@ class L2DomainScenarioTest1(ScenarioTest):
             'state_Disable': CONFIG.get('L2_ISOLATION_DOMAIN', 'state_Disable')
         })
 
-    def test_l2domain_scenario1(self):
+    def test_GA_l2domain_scenario1(self):
         ''' test scenario for L2 Domain CRUD operations'''
         call_scenario1(self)